refactor(assistant): split runtime concerns
